Reagrego funciones fuera de helpers al dict de helpers
