+++ ckanext/gobar_theme/utils/ckan_utils.py
@@ -1,12 +1,8 @@
-import pylons.config as config
 from pylons.controllers.util import abort
 
+from ckanext.gobar_theme.helpers import is_plugin_present
 
 TS_EXPLORER_PLUGIN = "seriestiempoarexplorer"
-
-def is_plugin_present(plugin_name):
-    plugins = config.get('ckan.plugins')
-    return plugin_name in plugins
 
 
 def plugin_or_404(plugin_name):
